[SR-377] Implement NSConcreteValue class
The NSConcreteValue class is a wrapper for arbitrary values that can be
described using Objective-C type notation. Note that owing to the lack of
factory initializers, a rather ugly workaround is used to make NSValue
behave as a class cluster. This should not have a performance impact in
practice as its init() is a NOOP and most NSValues will be NSNumbers.

NSConcreteValue does not yet support "special" values such as NSRects.
An implementation will be forthcoming in a separate commit.

[SR-377] NSConcreteValue should be a subclass of NSValue

[SR-377] cleanup NSConcreteValue, remove special cases

[SR-377]: NSConcreteValue: synchronize access to _cachedTypeInfo